requirements.txt
atomicwrites==1.1.5
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.8.2
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.2
entrypoints==0.3
et-xmlfile==1.0.
numpy==1.23.1
oauthlib==3.1.0
packaging==20.3
pandas==1.4.3
py==1.8.1
pytest==4.6.9
python-dotenv==0.20.0
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
SQLAlchemy==1.4.40
urllib3==1.25.8
